refactor(application): organize Auth feature into vertical slice structure with subfolders
