+++ CoreApp.Application/Features/Auth/Commands/Login/LoginCommand.cs
@@ -1,4 +1,4 @@
-﻿using CoreApp.Application.DTOs.Auth;
+﻿using CoreApp.Application.Features.Auth.DTOs;
 using MediatR;
 using System;
 using System.Collections.Generic;
