+++ CoreApp.Application/Features/Auth/Commands/Register/RegisterCommand.cs
@@ -1,4 +1,4 @@
-﻿using CoreApp.Application.DTOs.Auth;
+﻿using CoreApp.Application.Features.Auth.DTOs;
 using MediatR;
 
 namespace CoreApp.Application.Features.Auth.Commands.Register;
