+++ CoreApp.Application/Features/Auth/Commands/Register/RegisterCommandHandler.cs
@@ -1,5 +1,5 @@
-﻿using CoreApp.Application.DTOs.Auth;
-using CoreApp.Application.Interfaces.Auth;
+﻿using CoreApp.Application.Common.Interfaces.Auth;
+using CoreApp.Application.Features.Auth.DTOs;
 using MediatR;
 using System;
 using System.Collections.Generic;
